Completed MatrxF/D refactor
