+++ HellfireControl/src/HellfireControl/Math/Internal/Matrix_Common.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <HellfireControl/Core/Common.hpp>
+#include <HellfireControl/Math/Vector.hpp>
+
+#include <math.h>
+
+#if HC_USE_SIMD
+#include <intrin.h>
+#endif
