+++ HellfireControl/src/HellfireControl/Math/Internal/Quaternion_Common.hpp
@@ -1,0 +1,11 @@
+#pragma once
+
+#include <HellfireControl/Core/Common.hpp>
+#include <HellfireControl/Math/Vector.hpp>
+#include <HellfireControl/Math/Matrix.hpp>
+
+#include <math.h>
+
+#if HC_USE_SIMD
+#include <intrin.h>
+#endif
